chapter 10 all done

+++ modern_python_web_development_book/src/data/creature.py
@@ -1,5 +1,6 @@
-from .init import curs, conn
+from .init import curs, conn, IntegrityError
 from model.creature import Creature
+from errors import Missing, Duplicate
 
 curs.execute("""create table if not exists creature(
                 name text primary key,
@@ -10,7 +11,8 @@ curs.execute("""create table if not exists creature(
 
 def row_to_model(row: tuple) -> Creature:
     (name, description, country, area, aka) = row
-    return Creature(name, description, country, area, aka)
+    return Creature(name=name, description=description, 
+                    country=country, area=area, aka=aka)
 
 def model_to_dict(creature: Creature) -> dict:
     return creature.dict() 
@@ -19,7 +21,11 @@ def get_one(name: str) -> Creature:
     qry = "select * from creature where name=:name"
     params = {"name": name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(msg=f"Creature {name} not found")
 
 def get_all() -> list[Creature]:
     qry = "select * from creature"
@@ -27,14 +33,22 @@ def get_all() -> list[Creature]:
     return [row_to_model(row) for row in curs.fetchall()]
     
 def create(creature: Creature) -> Creature:
+    if not creature:
+        return None
     qry = """insert into creature values
         (:name, :description, :country, :area, :aka)"""
     params = model_to_dict(creature)
-    curs.execute(qry, params)
-    conn.commit()
+    try:
+        curs.execute(qry, params)
+        conn.commit()
+    except IntegrityError:
+        raise Duplicate(msg=
+            f"Creature {creature.name} already exists")
     return get_one(creature.name)
 
-def modify(creature: Creature) -> Creature:
+def modify(name: str, creature: Creature) -> Creature:
+    if not (name and creature): 
+        return None
     qry = """update creature
             set country=:country,
                 name=:name,
@@ -44,13 +58,20 @@ def modify(creature: Creature) -> Creature:
             where name=:name_orig"""
     params = model_to_dict(creature)
     params["name_orig"] = creature.name
-    _ = curs.execute(qry, params)
-    conn.commit()
-    return get_one(creature.name)
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        conn.commit()
+        return get_one(creature.name)
+    else:
+        raise Missing(msg=f"Creature {name} not found")
 
-def delete(creature: Creature) -> bool:
+def delete(name: str):
+    if not name: 
+        return False
     qry = "delete from creature where name = :name"
-    params = {"name": creature.name}
-    res = curs.execute(qry, params)
-    conn.commit()
-    return bool(res)
+    params = {"name": name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Creature {name} not found")
+    else:
+        conn.commit()
